requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.25
google-generativeai>=0.3.2
tavily-python>=0.2.8
python-dotenv>=1.0.0
aiohttp>=3.9.1